InteractiveEditingSession tests use standalone briefcase.
