+++ full-stack-tests/core/src/frontend/standalone/InteractiveEditingSession.test.ts
@@ -5,11 +5,11 @@
 import * as chai from "chai";
 import * as chaiAsPromised from "chai-as-promised";
 import * as path from "path";
-import { compareStrings, DbOpcode, Guid, Id64String, OpenMode, ProcessDetector } from "@bentley/bentleyjs-core";
+import { BeDuration, compareStrings, DbOpcode, Guid, Id64String, OpenMode, ProcessDetector } from "@bentley/bentleyjs-core";
 import { IModelJson, LineSegment3d, Point3d, Range3d, Transform, YawPitchRollAngles } from "@bentley/geometry-core";
-import { Code, ElementGeometryChange, ElementsChanged, IModelError, IModelWriteRpcInterface } from "@bentley/imodeljs-common";
+import { BatchType, Code, ElementGeometryChange, ElementsChanged, IModelError, IModelWriteRpcInterface } from "@bentley/imodeljs-common";
 import {
-  BriefcaseConnection, EditingFunctions, ElementEditor3d, InteractiveEditingSession,
+  BriefcaseConnection, EditingFunctions, ElementEditor3d, GeometricModel3dState, IModelTileTree, IModelTileTreeParams, InteractiveEditingSession, TileLoadPriority,
 } from "@bentley/imodeljs-frontend";
 import { ElectronApp } from "@bentley/electron-manager/lib/ElectronFrontend";
 
@@ -164,7 +164,6 @@ describe("InteractiveEditingSession", () => {
       const dictModelId = await imodel.models.getDictionaryModel();
       const category = await editing.categories.createAndInsertSpatialCategory(dictModelId, Guid.createValue(), { color: 0 });
       await imodel.saveChanges();
-      // await imodel.pushChanges("line 1"); // release locks
 
       // Begin an editing session.
       const session = await InteractiveEditingSession.begin(imodel);
@@ -236,8 +235,152 @@ describe("InteractiveEditingSession", () => {
 
       // ###TODO: No frontend API for testing undo/redo...
 
-      // await imodel.pushChanges(""); // release locks
       await session.end();
+      await editor.end();
+    });
+
+    it("updates state of tile trees", async () => {
+      const imodel = await openWritable();
+
+      // Initial geometric model contains one line element.
+      const editor = await ElementEditor3d.start(imodel);
+      const editing = new EditingFunctions(imodel);
+      const modelId = await editing.models.createAndInsertPhysicalModel(await editing.codes.makeModelCode(imodel.models.repositoryModelId, Guid.createValue()));
+      const dictModelId = await imodel.models.getDictionaryModel();
+      const category = await editing.categories.createAndInsertSpatialCategory(dictModelId, Guid.createValue(), { color: 0 });
+      const elem1 = await createLineElement(editor, modelId, category, makeLine(new Point3d(0, 0, 0), new Point3d(10, 0, 0)));
+      await imodel.saveChanges();
+
+      await imodel.models.load([modelId]);
+      const model = imodel.models.getLoaded(modelId) as GeometricModel3dState;
+      expect(model).not.to.be.undefined;
+      const location = Transform.createTranslationXYZ(0, 0, 2);
+
+      // NB: element ranges are minimum 1mm on each axis. Our line element is aligned to X axis.
+      const modelRange = await model.queryModelRange();
+      modelRange.low.y = modelRange.low.z = -0.0005;
+      modelRange.high.y = modelRange.high.z = 0.0005;
+      location.inverse()!.multiplyRange(modelRange, modelRange);
+
+      function createTileTree(): IModelTileTree {
+        const params: IModelTileTreeParams = {
+          id: "",
+          modelId,
+          iModel: imodel,
+          location,
+          priority: TileLoadPriority.Primary,
+          formatVersion: 14,
+          contentRange: modelRange.clone(),
+          options: {
+            allowInstancing: true,
+            edgesRequired: false,
+            batchType: BatchType.Primary,
+            is3d: true,
+          },
+          rootTile: {
+            contentId: "",
+            range: modelRange.toJSON(),
+            contentRange: modelRange.toJSON(),
+            maximumSize: 512,
+            isLeaf: true,
+          },
+        };
+
+        return new IModelTileTree(params);
+      }
+
+      const expectTreeState = async (tree: IModelTileTree | IModelTileTree[], expectedState: "static" | "interactive" | "dynamic" | "disposed", expectedHiddenElementCount: number, expectedRange: Range3d) => {
+        // ###TODO: After we switch from polling for native events, we should not need to wait for changed events to be fetched here...
+        const waitTime = 150;
+        await BeDuration.wait(waitTime);
+
+        const rangeTolerance = 0.0001;
+        const treeList = tree instanceof IModelTileTree ? [tree] : tree;
+        for (const t of treeList) {
+          expect(t.tileState).to.equal(expectedState);
+          expect(t.hiddenElements.length).to.equal(expectedHiddenElementCount);
+          expect(t.rootTile.range.isAlmostEqual(expectedRange, rangeTolerance)).to.be.true;
+          expect(t.rootTile.contentRange.isAlmostEqual(expectedRange, rangeTolerance)).to.be.true;
+          expect(t.contentRange!.isAlmostEqual(expectedRange, rangeTolerance)).to.be.true;
+        }
+      };
+
+      // No editing session currently active.
+      const tree1 = createTileTree();
+      expect(tree1.range.isAlmostEqual(modelRange)).to.be.true;
+      await expectTreeState(tree1, "static", 0, modelRange);
+
+      const tree0 = createTileTree();
+      tree0.dispose();
+      await expectTreeState(tree0, "disposed", 0, modelRange);
+
+      // Begin an editing session.
+      let session = await InteractiveEditingSession.begin(imodel);
+      const trees = [tree1, createTileTree()];
+      await expectTreeState(trees, "interactive", 0, modelRange);
+      await expectTreeState(tree0, "disposed", 0, modelRange);
+
+      // Insert a new element.
+      const elem2 = await createLineElement(editor, modelId, category, makeLine(new Point3d(0, 0, 0), new Point3d(-10, 0, 0)));
+      await imodel.saveChanges();
+
+      // Newly-inserted elements don't exist in tiles, therefore don't need to be hidden.
+      // ###TODO: Test changes to range and content range...
+      trees.push(createTileTree());
+      const range2 = modelRange.clone();
+      range2.low.x = -10;
+      await expectTreeState(tree0, "disposed", 0, modelRange);
+      await expectTreeState(trees, "dynamic", 0, range2);
+
+      // Modify an element.
+      await editor.startModifyingElements([elem1]);
+      await editor.applyTransform(Transform.createTranslationXYZ(0, 5, 0).toJSON());
+      await editor.write();
+      await imodel.saveChanges();
+
+      const range3 = range2.clone();
+      range3.high.y += 5;
+      trees.push(createTileTree());
+      await expectTreeState(tree0, "disposed", 0, modelRange);
+      await expectTreeState(trees, "dynamic", 1, range3);
+
+      // Delete the same element.
+      await deleteElements(imodel, [elem1]);
+      await imodel.saveChanges();
+      trees.push(createTileTree());
+      await expectTreeState(tree0, "disposed", 0, modelRange);
+      await expectTreeState(trees, "dynamic", 1, range2);
+
+      // Delete the other element.
+      await deleteElements(imodel, [elem2]);
+      await imodel.saveChanges();
+      trees.push(createTileTree());
+      await expectTreeState(tree0, "disposed", 0, modelRange);
+      await expectTreeState(trees, "dynamic", 2, modelRange);
+
+      // ###TODO: test undo/redo (no frontend API for that currently...)
+
+      // Terminate the session.
+      await session.end();
+      trees.push(createTileTree());
+      await expectTreeState(tree0, "disposed", 0, modelRange);
+      await expectTreeState(trees, "static", 0, modelRange);
+
+      // Restart session then terminate with no changes.
+      session = await InteractiveEditingSession.begin(imodel);
+      const tree2 = trees.pop()!;
+      await expectTreeState(tree0, "disposed", 0, modelRange);
+      await expectTreeState(trees, "interactive", 0, modelRange);
+      tree2.dispose();
+      await expectTreeState(tree2, "disposed", 0, modelRange);
+      await session.end();
+      await expectTreeState(trees, "static", 0, modelRange);
+
+      for (const tree of trees) {
+        tree.dispose();
+        await expectTreeState(tree, "disposed", 0, modelRange);
+      }
+
       await editor.end();
     });
   }
